@@ -9,15 +9,15 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.database.Cursor;
 import android.net.Uri;
-import android.support.v4.app.NavUtils;
-import android.support.v4.app.ShareCompat;
-import android.support.v4.content.LocalBroadcastManager;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
+import androidx.core.app.NavUtils;
+import androidx.core.app.ShareCompat;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.appcompat.widget.Toolbar;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.ProgressBar;
@@ -30,7 +30,6 @@ import com.apps.dbm.traveldbm.adapter.HotelAdapter;
 import com.apps.dbm.traveldbm.classes.Hotel;
 import com.apps.dbm.traveldbm.classes.Room;
 import com.apps.dbm.traveldbm.data.CollectionContract;
-import com.apps.dbm.traveldbm.service.HotelRequestService;
 import com.apps.dbm.traveldbm.widget.FavoriteAppWidgetProvider;
 
 import java.util.ArrayList;
