@@ -1,11 +1,10 @@
 package com.apps.dbm.traveldbm.service;
 
-import android.app.Application;
 import android.app.IntentService;
 import android.content.Intent;
 import android.net.Uri;
-import android.support.annotation.Nullable;
-import android.support.v4.content.LocalBroadcastManager;
+import androidx.annotation.Nullable;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 import android.util.Log;
 
 import com.android.volley.Request;
